+++ rgf/util.py
@@ -1,0 +1,222 @@
+import numpy as np
+import scipy.sparse as sp
+from sklearn.base import BaseEstimator, ClassifierMixin, RegressorMixin
+from sklearn.exceptions import NotFittedError
+from sklearn.externals import six
+from sklearn.utils.extmath import softmax
+from sklearn.utils.validation import check_array
+
+import rgf.sklearn
+
+
+def not_fitted_error_desc():
+    return "Estimator not fitted, call `fit` before exploiting the model."
+
+
+def sparse_savetxt(filename, input_array):
+    zip_func = six.moves.zip
+    if sp.isspmatrix_csr(input_array):
+        input_array = input_array.tocoo()
+    elif not sp.isspmatrix_coo(input_array):
+        input_array = input_array.tocsr().tocoo()
+    n_row = input_array.shape[0]
+    current_sample_row = 0
+    line = []
+    with open(filename, 'w') as fw:
+        fw.write('sparse {0:d}\n'.format(input_array.shape[-1]))
+        for i, j, v in zip_func(input_array.row, input_array.col, input_array.data):
+            if i == current_sample_row:
+                line.append('{0}:{1}'.format(j, v))
+            else:
+                fw.write(' '.join(line))
+                fw.write('\n' * (i - current_sample_row))
+                line = ['{0}:{1}'.format(j, v)]
+                current_sample_row = i
+        fw.write(' '.join(line))
+        fw.write('\n' * (n_row - i))
+
+
+def sigmoid(x):
+    return 1.0 / (1.0 + np.exp(-x))
+
+
+class RGFClassifierBase(BaseEstimator, ClassifierMixin):
+    @property
+    def estimators_(self):
+        """The collection of fitted sub-estimators when `fit` is performed."""
+        if self._estimators is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._estimators
+
+    @property
+    def classes_(self):
+        """The classes labels when `fit` is performed."""
+        if self._classes is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._classes
+
+    @property
+    def n_classes_(self):
+        """The number of classes when `fit` is performed."""
+        if self._n_classes is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._n_classes
+
+    @property
+    def n_features_(self):
+        """The number of features when `fit` is performed."""
+        if self._n_features is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._n_features
+
+    @property
+    def fitted_(self):
+        """Indicates whether `fit` is performed."""
+        if self._fitted is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._fitted
+
+    @property
+    def n_iter_(self):
+        """
+        Number of iterations of coordinate descent to optimize weights
+        used in model building process depending on the specified loss function.
+        """
+        if self._n_iter is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._n_iter
+
+    def predict_proba(self, X):
+        """
+        Predict class probabilities for X.
+
+        The predicted class probabilities of an input sample are computed.
+
+        Parameters
+        ----------
+        X : array-like or sparse matrix of shape = [n_samples, n_features]
+            The input samples.
+
+        Returns
+        -------
+        p : array of shape = [n_samples, n_classes].
+            The class probabilities of the input samples.
+            The order of the classes corresponds to that in the attribute classes_.
+        """
+        if self._fitted is None:
+            raise NotFittedError(not_fitted_error_desc())
+        X = check_array(X, accept_sparse=True)
+        n_features = X.shape[1]
+        if self._n_features != n_features:
+            raise ValueError("Number of features of the model must "
+                             "match the input. Model n_features is %s and "
+                             "input n_features is %s "
+                             % (self._n_features, n_features))
+        if self._n_classes == 2:
+            y = self._estimators[0].predict_proba(X)
+            y = sigmoid(y)
+            y = np.c_[y, 1 - y]
+        else:
+            y = np.zeros((X.shape[0], self._n_classes))
+            for i, clf in enumerate(self._estimators):
+                class_proba = clf.predict_proba(X)
+                y[:, i] = class_proba
+
+            # In honest, I don't understand which is better
+            # softmax or normalized sigmoid for calc probability.
+            if self.calc_prob == "sigmoid":
+                y = sigmoid(y)
+                normalizer = np.sum(y, axis=1)[:, np.newaxis]
+                normalizer[normalizer == 0.0] = 1.0
+                y /= normalizer
+            else:
+                y = softmax(y)
+        return y
+
+    def predict(self, X):
+        """
+        Predict class for X.
+
+        The predicted class of an input sample is computed.
+
+        Parameters
+        ----------
+        X : array-like or sparse matrix of shape = [n_samples, n_features]
+            The input samples.
+
+        Returns
+        -------
+        y : array of shape = [n_samples]
+            The predicted classes.
+        """
+        y = self.predict_proba(X)
+        y = np.argmax(y, axis=1)
+        return np.asarray(list(self._classes_map.values()))[np.searchsorted(list(self._classes_map.keys()), y)]
+
+    def cleanup(self):
+        """
+        Remove tempfiles used by this model.
+
+        Returns
+        -------
+        n_removed_files : int
+            Returns the number of removed files.
+        """
+        n_removed_files = 0
+        if self._estimators is not None:
+            for est in self._estimators:
+                n_removed_files += rgf.sklearn.cleanup_partial(est._file_prefix,
+                                                               remove_from_list=True)
+
+        # No more able to predict without refitting.
+        self._fitted = None
+        return n_removed_files
+
+
+class RGFRegressorBase(BaseEstimator, RegressorMixin):
+    @property
+    def n_features_(self):
+        """The number of features when `fit` is performed."""
+        if self._n_features is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._n_features
+
+    @property
+    def fitted_(self):
+        """Indicates whether `fit` is performed."""
+        if self._fitted is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._fitted
+
+    @property
+    def n_iter_(self):
+        """
+        Number of iterations of coordinate descent to optimize weights
+        used in model building process depending on the specified loss function.
+        """
+        if self._n_iter is None:
+            raise NotFittedError(not_fitted_error_desc())
+        else:
+            return self._n_iter
+
+    def cleanup(self):
+        """
+        Remove tempfiles used by this model.
+
+        Returns
+        -------
+        n_removed_files : int
+            Returns the number of removed files.
+        """
+        # No more able to predict without refitting.
+        self._fitted = None
+        return rgf.sklearn.cleanup_partial(self._file_prefix,
+                                           remove_from_list=True)
